Refactor: CustomImageUpload 리팩토링
